feat: compile weth9 and commit artifacts to vc
