+++ package.json
@@ -38,7 +38,7 @@
     "prettier": "^2.1.2",
     "prettier-plugin-solidity": "^1.0.0-alpha.59",
     "shelljs": "^0.8.4",
-    "solc": "0.5.12",
+    "solc": "0.4.18",
     "solhint": "^3.2.1",
     "solhint-plugin-prettier": "^0.0.5",
     "solidity-coverage": "^0.7.11",
