--- typechain/LibNote.d.ts
@@ -1,57 +1,0 @@
-/* Generated by ts-generator ver. 0.0.8 */
-/* tslint:disable */
-
-import {
-  ethers,
-  EventFilter,
-  Signer,
-  BigNumber,
-  BigNumberish,
-  PopulatedTransaction
-} from "ethers";
-import { Contract, ContractTransaction } from "@ethersproject/contracts";
-import { BytesLike } from "@ethersproject/bytes";
-import { Listener, Provider } from "@ethersproject/providers";
-import { FunctionFragment, EventFragment, Result } from "@ethersproject/abi";
-
-interface LibNoteInterface extends ethers.utils.Interface {
-  functions: {};
-
-  events: {
-    "LogNote(bytes4,address,bytes32,bytes32,bytes)": EventFragment;
-  };
-
-  getEvent(nameOrSignatureOrTopic: "LogNote"): EventFragment;
-}
-
-export class LibNote extends Contract {
-  connect(signerOrProvider: Signer | Provider | string): this;
-  attach(addressOrName: string): this;
-  deployed(): Promise<this>;
-
-  on(event: EventFilter | string, listener: Listener): this;
-  once(event: EventFilter | string, listener: Listener): this;
-  addListener(eventName: EventFilter | string, listener: Listener): this;
-  removeAllListeners(eventName: EventFilter | string): this;
-  removeListener(eventName: any, listener: Listener): this;
-
-  interface: LibNoteInterface;
-
-  functions: {};
-
-  callStatic: {};
-
-  filters: {
-    LogNote(
-      sig: BytesLike | null,
-      usr: string | null,
-      arg1: BytesLike | null,
-      arg2: BytesLike | null,
-      data: null
-    ): EventFilter;
-  };
-
-  estimateGas: {};
-
-  populateTransaction: {};
-}
