--- typechain/LibNoteFactory.ts
@@ -1,76 +1,0 @@
-/* Generated by ts-generator ver. 0.0.8 */
-/* tslint:disable */
-
-import { Signer } from "ethers";
-import { Provider, TransactionRequest } from "@ethersproject/providers";
-import { Contract, ContractFactory, Overrides } from "@ethersproject/contracts";
-
-import { LibNote } from "./LibNote";
-
-export class LibNoteFactory extends ContractFactory {
-  constructor(signer?: Signer) {
-    super(_abi, _bytecode, signer);
-  }
-
-  deploy(overrides?: Overrides): Promise<LibNote> {
-    return super.deploy(overrides || {}) as Promise<LibNote>;
-  }
-  getDeployTransaction(overrides?: Overrides): TransactionRequest {
-    return super.getDeployTransaction(overrides || {});
-  }
-  attach(address: string): LibNote {
-    return super.attach(address) as LibNote;
-  }
-  connect(signer: Signer): LibNoteFactory {
-    return super.connect(signer) as LibNoteFactory;
-  }
-  static connect(
-    address: string,
-    signerOrProvider: Signer | Provider
-  ): LibNote {
-    return new Contract(address, _abi, signerOrProvider) as LibNote;
-  }
-}
-
-const _abi = [
-  {
-    anonymous: true,
-    inputs: [
-      {
-        indexed: true,
-        internalType: "bytes4",
-        name: "sig",
-        type: "bytes4"
-      },
-      {
-        indexed: true,
-        internalType: "address",
-        name: "usr",
-        type: "address"
-      },
-      {
-        indexed: true,
-        internalType: "bytes32",
-        name: "arg1",
-        type: "bytes32"
-      },
-      {
-        indexed: true,
-        internalType: "bytes32",
-        name: "arg2",
-        type: "bytes32"
-      },
-      {
-        indexed: false,
-        internalType: "bytes",
-        name: "data",
-        type: "bytes"
-      }
-    ],
-    name: "LogNote",
-    type: "event"
-  }
-];
-
-const _bytecode =
-  "0x6080604052348015600f57600080fd5b50603e80601d6000396000f3fe6080604052600080fdfea265627a7a723158209e4bdc2b395aa9a08a26034d69f1bbaebb6804d2c496fac02a2716beb905751d64736f6c634300050c0032";
