+++ typechain/index.ts
@@ -1,4 +1,3 @@
 /* Generated by ts-generator ver. 0.0.8 */
 /* tslint:disable */
-export * from "./DaiFactory";
-export * from "./LibNoteFactory";
+export * from "./Weth9Factory";
